junit testing module, scripts for tests 1 and 2
